apps/wl1-python-django/pyproject.toml
[tool.poetry]
name = "wl1-python-django"
version = "0.1.0"
description = ""
authors = ["Evandro Pomatti <evandro.pomatti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pylint = "^2.14.2"
black = "^22.3.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
